app/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
audioop-lts==0.2.1
blinker==1.9.0
Brotli==1.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
discord.py==2.6.3
dotenv==0.9.9
Flask==3.1.2
frozenlist==1.7.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.6.3
mutagen==1.47.0
propcache==0.3.2
pycparser==2.22
pycryptodomex==3.23.0
PyNaCl==1.5.0
python-dotenv==1.1.1
redis==6.4.0
requests==2.32.5
spotipy==2.25.1
urllib3==2.5.0
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.20.1
yt-dlp>=2025.9.23
